+++ lumigator/python/mzai/evaluator/__init__.py
@@ -1,3 +1,1 @@
-from evaluator.entrypoint import Evaluator
 
-__all__ = ["Evaluator"]
